Update rand
Solaris 11 comes with shuf 8.16 which doesn't have "-r".  This was introduced in 8.22.  I considered putting in an "if uname = SunOS" check but decided it would be better to manage this behaviour based on version.  8.22 onwards will use "-r" and will be nice and fast.  Versions prior will use a shell loop, which does incur a performance penalty (0.16s for 10k numbers in the former, ~8s in the latter).  Testing also showed that shuf on Solaris likes its args provided in a particular order, so I've updated for that too.

This may get us towards modularising the code off into specific functions, which I tested as a response to finding the limitations of 'jot'