Update rand
Converted to a mostly functionalised script.  Built in the ability to zero-pad single digits, as well as to define whether or not printed elements can be repeatable.  At the moment this works for 'shuf' only, but will be pushed out across the other methods eventually.  I'm trying to think of a DRY way to achieve this.