New Fn_fill
After some testing, I'm pleased to publish the new and improved Fn_fill.  It doesn't necessarily appear to be faster at the smaller scale, but when we crank the testing up to a larger scale (e.g. 2000000 random ints), then the older method appears to be about 50% slower.  The new version is a bit more elegant and should eliminate any edge cases of sequential-pattern-creep as well.

I did want to look at a Fisher-Yates shuffle or even Reservoir Sampling, but the POSIX-esque nature of this script has tied my hands significantly.  No arrays makes those approaches tough, not to mention that even with arrays, they have their limits dependant on the shell that's used!