Correct method_jot for different versions of jot
As it turns out, different versions of jot generate different distribution types of random ints.  The previous version of rand worked fine on my Linux workstation using athena-jot, but not on a FBSD-10 VM.  This is covered brilliantly here: https://unix.stackexchange.com/a/241199.  This commit puts in a lo-fi solution, as there are no obvious ways to test versions or capabilities etc.