Update maxRand detection
I've tested `getconf LONG_BIT` across a number of shells and systems and it appears to be a very portable way to detect whether we're on a 32 or 64 bit system.  Previous attempts to automate maxRand/nMax have depended on what the shell itself reports, which is tricky when you've got, say, bash accepting 2^63-1 on a 32 bit system, and you try to get `shuf` to generate a number in that range.

If `getconf` doesn't work it fails over to 2^15-1, which is a commonly used range anyway.