Update xorshift to xorshift128+
After firing up a 32-bit VM and doing some testing, it looks like we can move the xorshift method up to a xorshift128+.  This is coupled with the arc4random_uniform style modulo debiasing correction that was used in the xorshift method.  I've also fixed the maxRand detection with specific handling for mksh.  As far as I know, it's the only potential target shell that has 32-bit arithmetic.